+++ app.js
@@ -1,6 +1,6 @@
 import express from "express";
 import cors from "cors";
-import { CLIENT_URL } from "./config/consts.js";
+import { CLIENT_URL, ROOMS_ROUTE } from "./config/consts.js";
 import userRouter from "./routes/User-route.js";
 import { USERS_ROUTE } from "./config/consts.js";
 import { router as roomsRouter } from "./routes/rooms/router.js";
@@ -17,7 +17,7 @@ export const initApp = () => {
 
   app.use(express.json());
   app.use(USERS_ROUTE, userRouter);
-  app.use("/api/rooms", roomsRouter);
+  app.use(ROOMS_ROUTE, roomsRouter);
 
   return app;
 };
